StackTrace filtering is done at collect time instead of display time. StackTraceText is sent with messagebroker instead of full StackTrace. Refactored classes away from Sql tab.
