+++ source/Glimpse.Ado/Message/CommandStackTraceMessage.cs
@@ -3,10 +3,10 @@ using System.Diagnostics;
 
 namespace Glimpse.Ado.Message
 {
-    public class CommandStackTraceMessage : AdoCommandPayloadMessage<StackTrace>
+    public class CommandStackTraceMessage : AdoCommandPayloadMessage<string>
     {
-        public CommandStackTraceMessage(Guid connectionId, Guid commandId, StackTrace stackTrace, bool isAsync = false)
-            : base(connectionId, commandId, stackTrace, isAsync)
+        public CommandStackTraceMessage(Guid connectionId, Guid commandId, string stackTraceText, bool isAsync = false)
+            : base(connectionId, commandId, stackTraceText, isAsync)
         {
         }
     }
