+++ source/Glimpse.Core/Glimpse.Core.csproj
@@ -109,6 +109,7 @@
     <Compile Include="Extensibility\JsonNetConverterDictionaryKeysAreNotPropertyNames.cs" />
     <Compile Include="Extensibility\InspectorContext.cs" />
     <Compile Include="Extensibility\RuntimePolicy.cs" />
+    <Compile Include="Extensibility\StackTraceFilter.cs" />
     <Compile Include="Extensibility\Subscriber.cs" />
     <Compile Include="Extensions\EnumExtensions.cs" />
     <Compile Include="Framework\AssemblyTypesResolver.cs" />
